Rename structures and correct EQU numbering.
A bunch of structures have accumulated in this directory over time, and
it got to the point where I was having a hard time keeping track of
which was which.  In particualar, I was surprised a few times by which
residue was in the active site (D vs E vs N).  So I gave all the
structures more consistent names to help make things more clear.

I also realized that most of the "lig_dimer" structures were missing
residue 251 (which used to be the EQU in the other monomer).  I must
have removed it for some reason (although I don't remember why now) but
neglected to fix the residue numbering.  So I fixed it just now.