Add the BuriedUnsatFilter
I used to have LoopModeler calculate this metric, but it calculates a
delta relative to -in:native, which is now *not* the wildtype structure.
That's not useful, so I added this metric to just calculate the absolute
number of buried unsats.  A benefit is that this makes it possible for
`compare_to_wildtype.sho` to visualize the buried unsats.